+++ src/Reservation.Domain/Entities/Categories/SubCategory.cs
@@ -6,7 +6,7 @@ public class SubCategory : BaseClass
     public string Title { get; set; }
     public string Description { get; set; }
     public string CoverImagePath { get; set; }
-    public int AveragePoint { get; set; }
+    public double AveragePoint { get; set; }
 
     // Parent Category
     public Category Category { get; set; }
